ImportanceNetwork now has its own file
