+++ binn/__init__.py
@@ -1,6 +1,6 @@
 from .binn import BINN
 from .logger import SuperLogger
-from .network import Network, ImportanceNetwork
+from .network import Network
+from .importance_network import ImportanceNetwork
 from .sklearn import BINNClassifier
-from .plot import *
 from .explainer import BINNExplainer
