+++ mkdocs.yml
@@ -14,6 +14,7 @@ nav:
       - reference/binn_ref.md
       - reference/net_ref.md
       - reference/explain_ref.md
+      - reference/import_ref.md
       - reference/sk_ref.md
 
 plugins:
